app/requirements.txt
attrs==19.3.0
case-conversion==2.1.0
certifi==2019.11.28
cffi==1.15.1
charset-normalizer==3.1.0
Deprecated==1.2.13
idna==3.4
inflection==0.3.1
marshmallow==3.19.0
marshmallow-jsonapi==0.24.0
packaging==23.0
pycparser==2.21
PyGithub==1.54.1
regex==2022.10.31
requests==2.28.2
ujson==3.0.0
urllib3==1.25.8
wrapt==1.11.2
